safari-600.7-branch/LayoutTests/svg/dom/SVGScriptElement/script-change-externalResourcesRequired-while-loading-expected.txt

Test changing externalResourcesRequired after inserting the element into the tree.
Test 1: PASS

Test 2: PASS

Test 3: PASS

Test 4: PASS
